Updated analysis report info to run a full check with dependencies (#49)
We were trying to limit the analysis time by using -m source-only
that will skip dependency analysis.  Then we learned there are a few
rules which will not properly run if we do this.  My current
understanding is that -m source-only is probably 'good enough' for our
needs, but to be thorough I'm omitting it for now so we can experiment
with the full amount of incidents possible.